requirement.txt
python-dotenv
torchvision
torch
dagster 
dagster-webserver
pandas
tele
fastapi
uvicorn
sqlalchemy
psycopg2-binary
opencv-python